move main.go to cmd/
